remove constants modules from rococo & kusama
