+++ runtime/mangata-kusama/src/lib.rs
@@ -65,7 +65,6 @@ use static_assertions::const_assert;
 pub use xcm::{latest::prelude::*, VersionedMultiLocation};
 
 pub use common_runtime::{currency::*, deposit, runtime_types, tokens, CallType};
-pub use constants::{fee::*, parachains::*};
 
 use mangata_support::traits::{
 	AssetRegistryApi, FeeLockTriggerTrait, PreValidateSwaps, ProofOfStakeRewardsApi,
@@ -85,7 +84,6 @@ use xyk_runtime_api::{RpcAmountsResult, XYKRpcResult};
 #[cfg(feature = "std")]
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
-pub mod constants;
 mod migration;
 mod weights;
 pub mod xcm_config;
@@ -420,7 +418,7 @@ impl pallet_transaction_payment_mangata::Config for Runtime {
 		FeeLock,
 	>;
 	type LengthToFee = cfg::pallet_transaction_payment_mangata::LengthToFee;
-	type WeightToFee = WeightToFee;
+	type WeightToFee = common_runtime::constants::fee::WeightToFee;
 	type FeeMultiplierUpdate = cfg::pallet_transaction_payment_mangata::FeeMultiplierUpdate;
 	type OperationalFeeMultiplier =
 		cfg::pallet_transaction_payment_mangata::OperationalFeeMultiplier;
